Made export code modular. Fixed exporting images bug.
When a set of images was exported, the image data for those images would then be replaced by the exported image name in storage, and then couldn't be exported again if they got deleted.

The image name was supposed to be added ALONGSIDE the image data, not in its place. This is corrected now.